+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Version 0.3.1
  * Cancelling the file selector for creating or restoring backups no longer causes an error (#3)
+ * Trying to save to an already existing blank text file will no longer cause an error (#5)
 
 #### Version 0.3.0
  * Added the ability to backup the mods for just one game